+++ dash/_callback.py
@@ -74,6 +74,7 @@ def callback(
     cancel: Optional[Union[List[Input], Input]] = None,
     manager: Optional[BaseBackgroundCallbackManager] = None,
     cache_args_to_ignore: Optional[list] = None,
+    cache_ignore_triggered=True,
     on_error: Optional[Callable[[Exception], Any]] = None,
     **_kwargs,
 ):
@@ -143,6 +144,9 @@ def callback(
             with keyword arguments (Input/State provided in a dict),
             this should be a list of argument names as strings. Otherwise,
             this should be a list of argument indices as integers.
+        :param cache_ignore_triggered:
+            Whether to ignore which inputs triggered the callback when creating
+            the cache.
         :param interval:
             Time to wait between the background callback update requests.
         :param on_error:
@@ -190,6 +194,8 @@ def callback(
 
         if cache_args_to_ignore:
             background_spec["cache_args_to_ignore"] = cache_args_to_ignore
+
+        background_spec["cache_ignore_triggered"] = cache_ignore_triggered
 
     return register_callback(
         callback_list,
@@ -404,11 +410,16 @@ def register_callback(
                 job_id = flask.request.args.get("job")
                 old_job = flask.request.args.getlist("oldJob")
 
+                cache_ignore_triggered = long.get("cache_ignore_triggered", True)
+
                 current_key = callback_manager.build_cache_key(
                     func,
                     # Inputs provided as dict is kwargs.
                     func_args if func_args else func_kwargs,
                     background.get("cache_args_to_ignore", []),
+                    None
+                    if cache_ignore_triggered
+                    else callback_ctx.get("triggered_inputs", []),
                 )
 
                 if old_job: